Fix full overlap elimination for inserts
In cases where there's a full overlap elimination for an insert
(e.g., when there's a delete followed by a insert that completely
overlaps the tail of the delete), the overlap is replaced by an equal
covering the overlap, and the insert should be changed to zero-length.

This adds a test for this overlap case for inserts (and the corresponding
overlap case for deletes), and it fixes a bug with the insert case.